+++ DotNet/Indieable.Sdk/IndieableClientOptions.cs
@@ -10,7 +10,7 @@ namespace IndieableSdk
     {
         public string BaseUrl { get; set; } = "https://indieable.com";
         public string PublicGameKey { get; set; } = "";
-        public string SdkVersion { get; set; } = "dotnet-0.4.0";
+        public string SdkVersion { get; set; } = "dotnet-0.4.1";
         public string BuildVersion { get; set; } = "";
         public string Platform { get; set; } = "";
         public string Environment { get; set; } = "production";